pyproject.toml
[tool.poetry]
name = "games-e-commerce-back"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.6"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.2.2"
Markdown = "^3.4.1"
django-filter = "^22.1"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
Pillow = "^9.4.0"
drf-spectacular = "^0.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
